requirements.txt
botocore
boto3
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-lambda
aws-cdk.core
aws-cdk.aws_dynamodb_global
aws-cdk.aws_dynamodb
